fix: refactor modal into its own file
